% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AP_pred.R
\name{ap_prediction}
\alias{ap_prediction}
\title{Atom Percent Prediction}
\usage{
ap_prediction(
  time,
  m0,
  n0,
  P,
  k,
  pool = "CH4",
  frac_P = P_default(pool),
  frac_k = k_default(pool)
)
}
\arguments{
\item{time}{vector of time values, numeric (e.g. days); first should be zero}

\item{m0}{pool size at time zero, as a volume}

\item{n0}{amount of heavy isotope at time zero, as a volume}

\item{P}{production rate, unit gas/unit time}

\item{k}{first-order rate constant for consumption, 1/unit time}

\item{pool}{Name of pool; see \code{\link{pdr_fractionation}}}

\item{frac_P}{Fractionation value for production; see \code{\link{pdr_fractionation}}}

\item{frac_k}{Fractionation value for consumption; see \code{\link{pdr_fractionation}}}
}
\value{
Returns a data frame with mt, nt, and AP (atom percent) predictions for each time step
}
\description{
Atom Percent Prediction
}
\note{
All combined, this is Eq. 11 from vFH2002 with a few modifications…
}
\author{
K.A. Morris & B. B-L
}
