% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_function.R
\name{cost_function}
\alias{cost_function}
\title{Cost Function}
\usage{
cost_function(
  params,
  time,
  m,
  n,
  m_prec,
  ap_prec,
  pool = "CH4",
  frac_P = P_default(pool),
  frac_k = k_default(pool),
  log_progress = NULL
)
}
\arguments{
\item{params}{Named vector holding optimizer-assigned values for P and k}

\item{time}{Vector of numeric time values (e.g. days); first should be zero}

\item{m}{Observed pool size (as a volume), same length as time}

\item{n}{Observed heavy isotope (as a volume), same length as time}

\item{m_prec}{Instrument precision for pool size, expressed as a standard deviation}

\item{ap_prec}{Instrument precision for atom percent, expressed as a standard deviation}

\item{pool}{Name of pool; see \code{\link{pdr_fractionation}}}

\item{frac_P}{Fractionation value for production; see \code{\link{pdr_fractionation}}}

\item{frac_k}{Fractionation value for consumption; see \code{\link{pdr_fractionation}}}

\item{log_progress}{An optional logging function}
}
\value{
Returns the sum of squares between predicted and observed m and AP
}
\description{
Cost Function
}
\note{
This is Eq. 14 from vFH2002 with a few modifications…
}
\examples{
m <- c(10, 8, 6, 5, 4, 3)
n <- c(1, 0.7, 0.6, 0.4, 0.3, 0.2)
cost_function(params = list(P = 0.5, k = 0.3), time = 0:5, m, n)
}
\author{
K.A. Morris & B. Bond-Lamberty
}
