% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{frac_k_default}
\alias{frac_k_default}
\title{Retrieve default k fractionation value for a pool}
\usage{
frac_k_default(pool)
}
\arguments{
\item{pool}{Name of pool, character}
}
\value{
The default entry for \code{pool} listed
in \code{\link{pdr_fractionation}}.
}
\description{
Retrieve default k fractionation value for a pool
}
\examples{
frac_k_default("CH4")
}
