% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{frac_P_default}
\alias{frac_P_default}
\title{Retrieve default P fractionation value for a pool}
\usage{
frac_P_default(pool)
}
\arguments{
\item{pool}{Name of pool, character}
}
\value{
The default entry for \code{pool} listed
in \code{\link{pdr_fractionation}}.
}
\description{
Retrieve default P fractionation value for a pool
}
\examples{
frac_P_default("CH4")
}
