% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{k_default}
\alias{k_default}
\title{Retrieve default k fractionation value for a gas}
\usage{
k_default(gas)
}
\arguments{
\item{gas}{Name of gas, character}
}
\value{
The default entry for \code{gas} listed
in \code{\link{pdr_fractionation}}.
}
\description{
Retrieve default k fractionation value for a gas
}
\examples{
k_default("CH4")
}
