% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Morris2023}
\alias{Morris2023}
\title{Example isotopic time series data from a methane pool experiment.}
\format{
\describe{
  \item{id}{Name of pool (gas or solid)}
  \item{time_days}{Fractionation value of production (P)}
  \item{cal12CH4ml}{Fractionation value of consumption (k)}
  \item{cal13CH4ml}{Default for this pool? Logical}
  \item{AP_obs}{Source paper or URL}
}
}
\usage{
Morris2023
}
\description{
Add description.
}
\keyword{datasets}
