% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{P_default}
\alias{P_default}
\title{Retrieve default P fractionation value for a gas}
\usage{
P_default(gas)
}
\arguments{
\item{gas}{Name of gas, character}
}
\value{
The default entry for \code{gas} listed
in \code{\link{pdr_fractionation}}.
}
\description{
Retrieve default P fractionation value for a gas
}
\examples{
P_default("CH4")
}
