% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pdr_fractionation}
\alias{pdr_fractionation}
\title{P and k fractionation values}
\format{
\describe{
  \item{Gas}{Gas}
  \item{frac_P}{Fractionation value of production (P)}
  \item{frac_k}{Fractionation value of consumption (k)}
  \item{Default}{Default for this gas? Logical}
  \item{Source}{Source paper or URL}
}
}
\usage{
pdr_fractionation
}
\description{
A compendium of possible production (P) and consumption (k) fractionation
values, by gas.
}
\keyword{datasets}
