% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\name{set_default_params}
\alias{set_default_params}
\title{Set default parameters for optimization}
\usage{
set_default_params(params, time, n, frac_k)
}
\arguments{
\item{params}{A named list of parameters}

\item{time}{Vector of numeric time values (e.g. days); first should be zero}

\item{n}{Observed heavy isotope (as a volume), same length as time}

\item{frac_k}{Fractionation value for consumption; see \code{\link{pdr_fractionation}}}
}
\value{
The \code{params} list with entries filled in as needed.
}
\description{
Set default parameters for optimization
}
