% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_function.R
\name{cost_function}
\alias{cost_function}
\title{Cost Function}
\usage{
cost_function(
  params,
  time,
  m,
  n,
  Nm,
  Nd,
  pool = "CH4",
  P,
  k,
  frac_P = P_default(pool),
  frac_k = k_default(pool),
  log_progress = NULL
)
}
\arguments{
\item{params}{Named list holding optimizer-assigned values for parameters}

\item{time}{Vector of numeric time values (e.g. days); first should be zero}

\item{m}{Observed pool size (as a volume), same length as time}

\item{n}{Observed heavy isotope (as a volume), same length as time}

\item{Nm}{Normalization factor for pool size, see Eq. 12}

\item{Nd}{Normalization factor for isotopic signature, see Eq. 13}

\item{pool}{Name of pool; see \code{\link{pdr_fractionation}}}

\item{P}{production rate, unit gas/unit time}

\item{k}{first-order rate constant for consumption, 1/unit time}

\item{frac_P}{Fractionation value for production; see \code{\link{pdr_fractionation}}}

\item{frac_k}{Fractionation value for consumption; see \code{\link{pdr_fractionation}}}

\item{log_progress}{An optional logging function}
}
\value{
Returns the sum of squares between predicted and observed m and AP
}
\description{
Cost Function
}
\note{
This is Eq. 14 from vFH2002 with a few modifications…
}
\examples{
m <- c(10, 8, 6, 5, 4, 3)
n <- c(1, 0.7, 0.6, 0.4, 0.3, 0.2)
Nm = m / 10
Nd = n / 10
cost_function(params = list(P = 0.5, k = 0.3), time = 0:5, m, n, Nm, Nd)
}
\author{
K.A. Morris & B. Bond-Lamberty
}
