% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_function.R
\name{cost_function}
\alias{cost_function}
\title{Cost Function}
\usage{
cost_function(
  params,
  time,
  m,
  n,
  Nm,
  Nd,
  pool = "CH4",
  frac_P = P_default(pool),
  frac_k = k_default(pool),
  log_progress = NULL
)
}
\arguments{
\item{params}{named vector holding optimizer-assigned values for P and k}

\item{time}{vector of time values, numeric (e.g. days); first should be zero}

\item{m}{observed pool size (as a volume), same length as time}

\item{n}{observed heavy isotope (as a volume), same length as time}

\item{Nm}{normalization factor for pool size, see Eq. 12}

\item{Nd}{normalization factor for isotopic signature, see Eq. 13}

\item{pool}{Name of pool; see \code{\link{pdr_fractionation}}}

\item{frac_P}{Fractionation value for production; see \code{\link{pdr_fractionation}}}

\item{frac_k}{Fractionation value for consumption; see \code{\link{pdr_fractionation}}}

\item{log_progress}{An optional logging function}
}
\value{
Returns the sum of squares between predicted and observed m and AP
}
\description{
Cost Function
}
\note{
This is Eq. 14 from vFH2002 with a few modifications…
}
\author{
K.A. Morris & B. B-L
}
