% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Morris2023}
\alias{Morris2023}
\title{Time series data from a methane pool dilution experiment.}
\format{
\describe{
  \item{id}{Sample ID}
  \item{time_days}{time in days between measurements, starting a 0}
  \item{cal12CH4ml}{ml of 12C-CH4 at each timestep}
  \item{cal13CH4ml}{ml of 13C-CH4 at each timestep}
  \item{AP_obs}{atom percent 13C-CH4 at each timestep}
}
}
\usage{
Morris2023
}
\description{
Sequential measurements of methane concentration and isotopic signature were taken using a Picarro G2920 with a Small Sample Introduction module.
This instrument provides gas concentrations in ppm and signatures in delta-13C, here we provide those data converted into volume of methane and atom percent.
}
\keyword{datasets}
