% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\name{optimize_pk}
\alias{optimize_pk}
\title{Optimize P (production) and k (consumption) for a pool dilution experiment}
\usage{
optimize_pk(
  time,
  m,
  n,
  m_prec,
  ap_prec,
  params,
  pool = "CH4",
  frac_P = P_default(pool),
  frac_k = k_default(pool),
  cost_fn = cost_function,
  prediction_fn = ap_prediction
)
}
\arguments{
\item{time}{Vector of numeric time values (e.g. days); first should be zero}

\item{m}{Observed pool size (as a volume), same length as time}

\item{n}{Observed heavy isotope (as a volume), same length as time}

\item{m_prec}{Instrument precision for pool size, expressed as a standard deviation}

\item{ap_prec}{Instrument precision for atom percent, expressed as a standard deviation}

\item{params}{Parameters TODO}

\item{pool}{Name of pool; see \code{\link{pdr_fractionation}}}

\item{frac_P}{Fractionation value for production; see \code{\link{pdr_fractionation}}}

\item{frac_k}{Fractionation value for consumption; see \code{\link{pdr_fractionation}}}

\item{cost_fn}{Cost function to use; the default is \code{\link{cost_function}}}

\item{prediction_fn}{Prediction function that the cost function will use;
the default is \code{\link{ap_prediction}}}
}
\value{
The output of \code{\link{optim}}.
}
\description{
Optimize P (production) and k (consumption) for a pool dilution experiment
}
\examples{
m <- c(10, 8, 6, 5, 4, 3)
n <- c(1, 0.7, 0.6, 0.4, 0.3, 0.2)
optimize_pk(time = 0:5, m, n, m_prec = 0.001, ap_prec = 1, params = list(P0 = 0.5))
}
