% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\name{set_default_params}
\alias{set_default_params}
\title{Set default parameters for optimization}
\usage{
set_default_params(other_params)
}
\arguments{
\item{other_params}{A named list of parameters}
}
\value{
The \code{params} list with entries filled in as needed.
}
\description{
Set default parameters for optimization
}
